accel/amdxdna: move ctx status query to amdxdna_ctx_status.c
Extract amdxdna_get_hwctx_status(), amdxdna_query_ctx_status_array(),
and amdxdna_query_ctx_status_by_id() along with their supporting
helpers and walk struct from aie.c into a dedicated
amdxdna_ctx_status.c / amdxdna_ctx_status.h pair. Update aie2_pci.c
and aie4_pci.c to include the new header, and add the new object to
Kbuild.
