+++ drivers/accel/amdxdna/aie.h
@@ -180,12 +180,6 @@ int amdxdna_get_aie_status(struct aie_device *aie,
 			   struct amdxdna_drm_get_info *args);
 int amdxdna_get_telemetry(struct aie_device *aie, struct amdxdna_client *client,
 			  struct amdxdna_drm_get_info *args);
-int amdxdna_get_hwctx_status(struct aie_device *aie, struct amdxdna_client *client,
-			     struct amdxdna_drm_get_info *args);
-int amdxdna_query_ctx_status_array(struct aie_device *aie, struct amdxdna_client *client,
-				   struct amdxdna_drm_get_array *args);
-int amdxdna_query_ctx_status_by_id(struct aie_device *aie, struct amdxdna_client *client,
-				   struct amdxdna_drm_get_array *args);
 int amdxdna_get_force_preempt_state(struct aie_device *aie, struct amdxdna_drm_get_info *args);
 int amdxdna_get_frame_boundary_preempt_state(struct aie_device *aie,
 					     struct amdxdna_drm_get_info *args);
